+++ packages/orq-rc/RELEASES.md
@@ -12249,3 +12249,13 @@ Based on:
 - [typescript v4.15.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.36 - packages/orq-rc
+
+## 2026-09-10 14:50:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.37 - packages/orq-rc